Refactor util functions and add unit tests
