+++ components/file_upload/file_upload.jsx
@@ -16,7 +16,7 @@ import {
     isIosChrome,
     isMobileApp,
 } from 'utils/user_agent.jsx';
-import {hasTable} from 'utils/post_utils.jsx';
+import {getTable} from 'utils/paste.jsx';
 import {
     clearFileInput,
     cmdOrCtrlPressed,
@@ -414,7 +414,7 @@ export default class FileUpload extends PureComponent {
     pasteUpload = (e) => {
         const {formatMessage} = this.context.intl;
 
-        if (!e.clipboardData || !e.clipboardData.items || hasTable(e.clipboardData)) {
+        if (!e.clipboardData || !e.clipboardData.items || getTable(e.clipboardData)) {
             return;
         }
